refactor play command

+++ cogs/MusicPlayer.py
@@ -35,56 +35,37 @@ class MusicPlayer(commands.Cog):
                 option = int(music[-1])
                 music = music[:-3]
         except Exception as e:
-            pass
-
-        if option is None:
-            song_there = os.path.isfile("song.mp3")
-            try:
-                if song_there:
-                    os.remove("song.mp3")
-                    await ctx.send("Preparing song")
-            except PermissionError:
-                await ctx.send("Song is playing now. Your song will be in queue")
-                global video_choice
-                if music.isdigit():
-                    if int(music) > len(video_choice):
-                        await ctx.send("Wrong option")
-                        return
-                    else:
-                        music = video_choice[int(music) - 1]
-
-                video_choice = {}
-                try:
-                    title = music_title(music)
-                    song_queue.append([title, music])
-                    await ctx.send("{} was placed in queue".format(title))
-                except Exception as e:
-                    print(e)
-                    videos = search(music)
-                    embed = discord.Embed(title="Select song", color=0x0ff0ff)
-                    for i in range(len(videos)):
-                        embed.add_field(name="** **", value="**{}**. {} | {}".format(i + 1, videos[i][3], videos[i][5]),
-                                        inline=False)
-                        video_choice[i] = videos[i][2]
-                    await ctx.send(embed=embed)
-                return
-
+            print(e)
             if music.isdigit():
                 if int(music) > len(video_choice):
                     await ctx.send("Wrong option")
                     return
                 else:
-                    music = video_choice[int(music)-1]
+                    music = video_choice[int(music) - 1]
 
-            video_choice = {}
-            try:
-                voice = get(self.bot.voice_clients, guild=ctx.guild)
+        song_play = True
+        song_there = os.path.isfile("song.mp3")
+        try:
+            if song_there:
+                os.remove("song.mp3")
+                await ctx.send("Preparing song")
+            song_play = False
+        except PermissionError:
+            await ctx.send("Song is playing now. Your song will be in queue")
+
+        try:
+            if song_play:
+                title = music_title(music)
+                song_queue.append([title, music])
+                await ctx.send("{} was placed in queue".format(title))
+            else:
                 title = download(music)
-                await ctx.send("Downloading music")
                 await ctx.send("{} is playing now".format(title))
                 voice.play(discord.FFmpegPCMAudio("song.mp3"))
-            except Exception as e:
-                print(e)
+        except Exception as e:
+            print(e)
+            if option is None:
+                video_choice = {}
                 videos = search(music)
                 embed = discord.Embed(title="Select song", color=0x0ff0ff)
                 for i in range(len(videos)):
@@ -92,31 +73,26 @@ class MusicPlayer(commands.Cog):
                                     inline=False)
                     video_choice[i] = videos[i][2]
                 await ctx.send(embed=embed)
-        else:
-            videos = search(music)
-            for i in range(len(videos)):
-                video_choice[i] = videos[i][2]
-
-            if option > len(video_choice):
-                await ctx.send("Wrong option")
-                return
             else:
-                music = video_choice[option - 1]
+                videos = search(music)
+                for i in range(len(videos)):
+                    video_choice[i] = videos[i][2]
 
-            song_there = os.path.isfile("song.mp3")
-            try:
-                if song_there:
-                    os.remove("song.mp3")
-                    await ctx.send("Preparing song")
-                await ctx.send("Downloading music")
-                title = download(music)
-                await ctx.send("{} is playing now".format(title))
-                voice.play(discord.FFmpegPCMAudio("song.mp3"))
-            except PermissionError:
-                await ctx.send("Song is playing now. Your song will be in queue")
-                title = music_title(music)
-                song_queue.append([title, music])
-                await ctx.send("{} was placed in queue".format(title))
+                if option > len(video_choice):
+                    await ctx.send("Wrong option")
+                    return
+                else:
+                    music = video_choice[option - 1]
+
+                if song_play:
+                    title = music_title(music)
+                    song_queue.append([title, music])
+                    await ctx.send("{} was placed in queue".format(title))
+                else:
+                    title = download(music)
+                    await ctx.send("{} is playing now".format(title))
+                    voice.play(discord.FFmpegPCMAudio("song.mp3"))
+        return
 
     @commands.command(pass_context=True, aliases=['pa', 'p_'])
     async def pause(self, ctx):
@@ -189,24 +165,31 @@ class MusicPlayer(commands.Cog):
         if str(ctx.channel) != 'music':
             return
 
-        voice = get(self.bot.voice_clients, guild=ctx.guild)
+        if len(song_queue) > 0:
+            voice = get(self.bot.voice_clients, guild=ctx.guild)
 
-        if voice and voice.is_playing():
-            voice.stop()
+            if voice and voice.is_playing():
+                voice.stop()
+                sleep(0.5)
 
-        sleep(0.5)
-        song_there = os.path.isfile("song.mp3")
-        if song_there:
-            os.remove("song.mp3")
-            await ctx.send("Preparing song")
+            song_there = os.path.isfile("song.mp3")
+            if song_there:
+                os.remove("song.mp3")
+                await ctx.send("Preparing song")
 
-        await ctx.send("Next Song")
-        await ctx.send("Downloading music")
-        music = song_queue[0]
-        song_queue.remove(music)
-        title = download(music[1])
-        voice.play(discord.FFmpegPCMAudio("song.mp3"))
-        await ctx.send("{} is playing now".format(title))
+            await ctx.send("Next Song")
+            await ctx.send("Downloading music")
+            music = song_queue[0]
+            song_queue.remove(music)
+            title = download(music[1])
+            voice.play(discord.FFmpegPCMAudio("song.mp3"))
+            await ctx.send("{} is playing now".format(title))
+        else:
+            await ctx.send("No song in queue")
+
+    def play_next(self, ctx):
+        self.next_song(ctx)
+        return
 
 
 def setup(bot):
